+++ spec/support/devise.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+RSpec.configure do |config|
+  config.include Warden::Test::Helpers
+  config.after :each do
+    Warden.test_reset!
+  end
+end