+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Possibility to configure development DB using ENV variables (@mkasztelnik)
 - Bootstrap JS and font awesome regular (@mszostak)
 - Add styling for services, improve global styles (@kosmidma)
+- User affiliations (@mkasztelnik)
 
 ### Changed
 - Upgrade Sprockets gem to avoid CVE-2018-3760 vulnerability (@mkasztelnik)